Added policy, minor fixes
Added policy for tasks, changed unsignedBigInt to foreignId in task migration, added prefix for task routes, moved geting task from page controller to task controller

+++ app/Http/Controllers/TaskController.php
@@ -4,34 +4,49 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\StoreTaskRequest;
 use Illuminate\Http\Request;
+use Gate;
 use App\Models\Task;
 use Auth;
 
 
 class TaskController extends Controller
 {
+    public function index()
+    {
+        $tasks = Auth::user()->tasks()->get();
+
+        return view('todo', [
+            'tasks' => $tasks,
+        ]);
+    }
+
     public function create(StoreTaskRequest $request)
-    {        
-        Task::create([
+    {
+        Auth::user()->tasks()->create([
             'title' => $request->validated()['title'],
             'user_id' => Auth::user()->id,
         ]);
-        
+
         return redirect()->back()->with('status', 'Task added');
     }
 
     public function delete(Task $task)
     {
+        Gate::authorize('delete', $task);
+
         $task->delete();
 
         return redirect()->back()->with('status', 'Task deleted');
     }
 
-    public function edit(Request $request,  Task $task)
+    public function edit(Request $request, Task $task)
     {
-        $task->title = $request->title;
+        Gate::authorize('update', $task);
 
-        $task->save();
+
+        $task->update([
+            'title' => $request->title,
+        ]);
 
         return redirect()->back()->with('status', 'Task updated');
     }
